Training Problem
The training problem is the most relevant type for the user. The rest of
the commits focus on supporting this type.

Models.jl contains a list of models that are available for the user for
reducing the size of the problem.

TrainingData.jl contains types for all the supported data that can be
used for this reduction purpose.

Integrators.jl contains a list of integrators and would eventually
contain interfaces for resolving the problem in the reduced space.